refactored scroll to top button as its own shared component. added skeleton of toolbar
