+++ app/academic-writing/mockData.js
@@ -1,0 +1,78 @@
+export const mockData = [
+  {
+    title: 'Title',
+    publicationYear: 2021,
+    description:
+      'Lorem ipsum dolor Lorem ipsum dolor s Lorem ipsum Lorem ipsum dolor s Lorem ipsum dolor s dolor s Lorem ipsum dolor s sfldklkmvlkmdflvmit  joidfjio ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 1990,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 1990,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 1989,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 2007,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 2012,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 1997,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 2001,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 1990,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 2020,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 1995,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 1990,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 1990,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 1990,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+  {
+    title: 'Title',
+    publicationYear: 2008,
+    description: 'Lorem ipsum dolor sit ame',
+  },
+];
