+++ app/shared components/Buttons/ScrollToTop/ScrollToTopButton.module.scss
@@ -1,0 +1,30 @@
+@import 'colors';
+@import 'breakpoints';
+@import 'variables';
+
+.toTop {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 35px;
+  width: 35px;
+  position: fixed;
+  bottom: 20px;
+  right: 20px;
+  background-color: $meadow-trail;
+  border: none;
+  border-radius: $icon-button-border;
+  cursor: pointer;
+
+  img {
+    filter: $white-filter;
+  }
+
+  @media (max-width: $breakpoint-sm) {
+    height: 40px;
+    width: 40px;
+    right: 0;
+    bottom: 70px;
+    border-radius: $icon-button-border 0 0 $icon-button-border;
+  }
+}
